hw12-books-spring-web-security/src/main/java/ru/otus/books/services/UserServiceImpl.java
package ru.otus.books.services;

import org.springframework.stereotype.Service;
import ru.otus.books.domain.User;
import ru.otus.books.repositories.UserRepository;

import java.util.Optional;

@Service
public class UserServiceImpl implements UserService {

    private final UserRepository repository;

    public UserServiceImpl(UserRepository repository) {
        this.repository = repository;
    }

    @Override
    public Optional<User> findByUsername(String username) {
        return repository.findByUsername(username);
    }
}
